Display 'N/A' for peak BW instead of '0' when it's not available
The display of 0 peak BW is confusing, especially when the user does not
choose to run without peak-bw explicitly, but for example runs with
'-D'. In such cases peak-bw calculatiuon is disabled internally, and the
zero result is confusing, and makes the user assume that something went
wrong during the run. If the value is not calculated, it shouldn't be
numeric.

Signed-off-by: Leonid Bloch <lb.workbox@gmail.com>